Refactor Cardbutton to own file
